requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
deepface>=0.0.90
tensorflow>=2.13.0
